Significantly reworked function handling
Functions now each get their own specific KExpression. This involved a
lot of additions and some rewriting in the parser and interpreter, but
it's a lot cleaner now.

Also updated the front-facing interpreter to output the type of the
KExpression instead of the parsed value. This can be adjusted with
command-line options.